+++ changelog.md
@@ -1,6 +1,10 @@
 # Changelog
 
-## [Unreleased]
+## [3.3.0] - 25.08.2026
+
+### Added
+
+- Added option to save collapsed sections in user meta
 
 ### Changed
 
@@ -9,6 +13,7 @@
 ### Fixed
 
 - Fixed to hide save for tabs with sub-tabs in classic view
+- Fixed handling of default sub-tabs and their ordering in DataView
 
 ## [3.2.1] - 24.08.2026